general "compressedsegment" class
